Re-arrange files to fix tests.
